Only adjust device.brightness in response to confirmed attempts to set brightness.
This change deals with the  problem that level changes to the device are asynchronous
and non-instantaneous.

If we store the brightness level before we reach the target brightness, the brightness will be reset to the wrong level across an on/off cycle.

So, instead, we only update device.brightness when it settles down following a change.

Signed-off-by: Jon Seymour <jon@ninjablocks.com>